feat(rich-text-editor): refactor link plugin to new API
