--- packages/picasso-rich-text-editor/src/LexicalLinkPlugin/LexicalLinkPlugin.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import { LinkPlugin } from '@lexical/react/LexicalLinkPlugin'
-
-export const LexicalLinkPlugin = () => {
-  return <LinkPlugin />
-}
