--- packages/picasso-rich-text-editor/src/LexicalLinkPlugin/index.ts
@@ -1,2 +1,0 @@
-export * from './LexicalLinkPluginButton'
-export * from './LexicalLinkPlugin'
