+++ packages/picasso-rich-text-editor/src/LinkPlugin/index.ts
@@ -1,0 +1,2 @@
+export { LinkPluginButton } from './LinkPluginButton'
+export * from './LinkPlugin'
